+++ doc/source/changelog.rst
@@ -6,9 +6,8 @@ Next Release
 ============
 
 **Features Added**
-* Added ``replace`` :class:`~montepy.surfaces.half_space.HalfSpace` to swap dividers in a cell geometry tree, and ``__iter__`` to traverse geometry leaves (:issue:`737`).
 
-
+* Added ``HalfSpace.replace`` to swap dividers in a cell geometry tree, and ``HalfSpace.__iter__`` to traverse geometry leaves (:issue:`737`).
 1.4 releases
 ============